Disable strict in bundles
Because the code is not designed with strict in mind and rollup automatically adds "use strict;" at the top of generated bundles, we get all sorts of issues on runtime. One error i've encountered is in `mxCellRenderer.prototype.isTextShapeInvalid` where in function `check`, `result` is not defined in strict mode.

This PR just disables `use strict;` in rollup, as fixing the underlying code could add to many regressions.